src/main/java/wooteco/subway/domain/Path.java
package wooteco.subway.domain;

import java.util.Collections;
import java.util.List;
import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
import org.jgrapht.graph.DefaultWeightedEdge;
import org.jgrapht.graph.WeightedMultigraph;

public class Path {

    private static final WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);

    private List<Station> stations;

    public Path(List<Section> sections, Station departure, Station arrival) {
        for (Section section : sections) {
            final Station upStation = section.getUpStation();
            final Station downStation = section.getDownStation();

            graph.addVertex(upStation);
            graph.addVertex(downStation);
            graph.setEdgeWeight(graph.addEdge(upStation, downStation), section.getDistance());
        }

        DijkstraShortestPath<Station, DefaultWeightedEdge> dijkstraShortestPath = new DijkstraShortestPath<>(graph);
        this.stations = dijkstraShortestPath.getPath(departure, arrival).getVertexList();
    }

    public List<Station> getStations() {
        return Collections.unmodifiableList(stations);
    }
}
